upgrade the depreciated np.float to np.float32

+++ motion_editing/fabrik_chain.py
@@ -99,13 +99,13 @@ class FABRIKBone(object):
     def __init__(self, name, child):
         self.name = name
         self.child = child
-        self.position = np.array([0, 0, 0], np.float) # position of joint
+        self.position = np.array([0, 0, 0], np.float32) # position of joint
         self.length = 0
         self.is_root = False
         self.is_leaf = False
 
 
-ROOT_OFFSET = np.array([0,0,0], np.float)
+ROOT_OFFSET = np.array([0,0,0], np.float32)
 
 class FABRIKChain(object):
     def __init__(self, skeleton, bones, node_order, tolerance=0.01, delta_tolerance=0.0001, max_iter=500, frame_offset=3, root_offset=ROOT_OFFSET, activate_constraints=False, visualize=False):
@@ -514,7 +514,7 @@ class FABRIKChain(object):
     def get_child_offset(self, node, child_node):
         """
         """
-        actual_offset = np.array([0, 0, 0], np.float)
+        actual_offset = np.array([0, 0, 0], np.float32)
         while node is not None and self.skeleton.nodes[node].children[0].node_name != child_node:
             local_offset = np.array(self.skeleton.nodes[node].children[0].offset)
             local_offset /= np.linalg.norm(local_offset)
